Updates styling and migrates to React Router v6
Updates the application's styling to use a new color scheme and improve the user interface.

Migrates the application to React Router v6, streamlining routing logic and improving overall maintainability.
Removes the Navbar from the App component and simplifies the App component to only include the routing logic.
Updates the styling of the checkbox.
Updates the Vectorpedia title description.